fix(e2e): reduce e2e flakiness when a deployment pod restarts
