+++ e2e/kube/debug.go
@@ -23,7 +23,6 @@ import (
 
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/fields"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/client-go/rest"
@@ -138,18 +137,4 @@ func workloadPods(ctx context.Context, kubeClient client.Client, o client.Object
 	default:
 		return nil, errors.New("invalid object type")
 	}
-}
-
-func podsFromSelector(ctx context.Context, kubeClient client.Client, ps *metav1.LabelSelector) ([]corev1.Pod, error) {
-	selector, err := metav1.LabelSelectorAsSelector(ps)
-	if err != nil {
-		return nil, err
-	}
-	podList := &corev1.PodList{}
-	if err := kubeClient.List(ctx, podList, &client.ListOptions{
-		LabelSelector: selector,
-	}); err != nil {
-		return nil, err
-	}
-	return podList.Items, nil
 }
