+++ e2e/ruler_test.go
@@ -719,6 +719,7 @@ func ruleEvaluatorPod(ctx context.Context, kubeClient client.Client, namespace s
 	if err != nil {
 		return nil, err
 	}
+	podList = kube.PodsReady(podList)
 	if len(podList) != 1 {
 		return nil, fmt.Errorf("expected 1 pod, found %d", len(podList))
 	}
